Add files via upload
In this game there are 4 colored button. If we press any key then game will be started and blink a random color. we have to tap that blink color then we will move on next level and again a random color will blink , then we we have to tap previous blink color  and current blink color if we tap right color then we will move on next level otherwise we will lose and we can again restart this game by just clicking any key.